Make it so the KYCWall can be re-used for Transfer Balance flow
